# docs/api-reference/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bodhveda Notifications API",
        "version": "1.0",
        "description": "API for sending notifications to recipients or broadcasting to all."
    },
    "paths": {
        "/notifications/send": {
            "post": {
                "summary": "Send a notification (direct or broadcast)",
                "operationId": "sendNotification",
                "tags": [
                    "Notifications"
                ],
                "description": "Start sending notifications through Bodhveda API.\n\nA send names an optional `target` (`channel`/`topic`/`event`) and carries a `payload` (in-app), an `email` block, or both. The mediums it asks for are the mediums it carries.\n\n### Targets and your catalog\n\nBy default you can send **any** target, cataloged or not. Cataloging a target — creating a project preference for it — is what puts it on your recipients' preference screen and lets `email` resolve to deliverable; an uncataloged target still sends in-app, but no recipient can mute it.\n\nIf the project has **strict targets** turned on (a per-project setting, **off by default**), this endpoint instead rejects a send whose `(target, medium)` has no matching catalog entry — see the `400` below. A `topic: any` catalog entry satisfies the check for every concrete topic beneath it, so one entry covers an unbounded set of runtime-generated topics; you never need one entry per resource id. A send carrying no `target` at all is never checked.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendNotificationPayload"
                            }
                        }
                    }
                },
                "x-codeSamples": [
                    {
                        "lang": "curl",
                        "label": "cURL",
                        "source": "curl -X POST https://api.bodhveda.com/notifications/send \\\n  -H \"Authorization: Bearer bv_xxxxxxxxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"recipient_id\": \"recipient_123\",\n    \"target\": {\n      \"channel\": \"posts\",\n      \"topic\": \"post_id_123\",\n      \"event\": \"new_comment\"\n    },\n    \"payload\": {\n      \"title\": \"Elon, Zuck and others commented on your post.\",\n      \"post_url\": \"/posts/post_id_123#comments\"\n    }\n  }' "
                    },
                    {
                        "lang": "curl",
                        "label": "cURL — with email",
                        "source": "# Direct send fanned out to inbox + email.\n# The `email` block makes email eligible (direct sends only).\ncurl -X POST https://api.bodhveda.com/notifications/send \\\n  -H \"Authorization: Bearer bv_xxxxxxxxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"recipient_id\": \"recipient_123\",\n    \"target\": {\n      \"channel\": \"digest\",\n      \"topic\": \"none\",\n      \"event\": \"sent\"\n    },\n    \"payload\": {\n      \"title\": \"Your daily digest is ready.\"\n    },\n    \"email\": {\n      \"subject\": \"Your daily digest\",\n      \"html\": \"<h1>Your daily digest</h1><p>3 new follow-ups today.</p>\",\n      \"text\": \"Your daily digest\\n3 new follow-ups today.\"\n    }\n  }' "
                    },
                    {
                        "lang": "javascript",
                        "label": "JavaScript",
                        "source": "import { Bodhveda } from \"@bodhveda/js\";\n\nconst bodhveda = new Bodhveda(\"bv_xxxxxxxxx\");\n\nconst res = await bodhveda.notifications.send({\n    recipient_id: \"recipient_123\",\n    target: { channel: \"posts\", topic: \"post_id_123\", event: \"new_comment\" },\n    payload: {\n        title: \"Elon, Zuck and others commented on your post.\",\n        post_url: \"/posts/post_id_123#comments\",\n    },\n});\n// res.notification (direct) or res.broadcast, plus res.deliveries for email."
                    },
                    {
                        "lang": "go",
                        "label": "Go",
                        "source": "import (\n    \"context\"\n    \"encoding/json\"\n\n    bodhveda \"github.com/MudgalLabs/bodhveda/sdk/go\"\n)\n\nctx := context.Background()\nclient := bodhveda.NewClient(\"bv_xxxxxxxxx\", nil)\n\nrecipientID := \"recipient_123\"\npayload, _ := json.Marshal(map[string]any{\n    \"title\":    \"Elon, Zuck and others commented on your post.\",\n    \"post_url\": \"/posts/post_id_123#comments\",\n})\nres, _ := client.Notifications.Send(ctx, &bodhveda.SendNotificationRequest{\n    RecipientID: &recipientID,\n    Target:      &bodhveda.Target{Channel: \"posts\", Topic: \"post_id_123\", Event: \"new_comment\"},\n    Payload:     payload,\n})\n// res.Notification (direct) or res.Broadcast, plus res.Deliveries for email."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification response",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Direct notification": {
                                        "summary": "Direct Notification Response",
                                        "value": {
                                            "message": "Direct notification queued for delivery to recipient recipient_123.",
                                            "data": {
                                                "notification": {
                                                    "id": 42069,
                                                    "recipient_id": "recipient_123",
                                                    "payload": {
                                                        "title": "Elon, Zuck and others commented on your post.",
                                                        "post_url": "/posts/post_id_123#comments"
                                                    },
                                                    "broadcast_id": null,
                                                    "target": {
                                                        "channel": "posts",
                                                        "topic": "post_id_123",
                                                        "event": "new_comment"
                                                    },
                                                    "read": false,
                                                    "opened": false,
                                                    "created_at": "2025-11-07T05:31:56Z",
                                                    "updated_at": "2025-11-07T05:31:56Z",
                                                    "status": "enqueued"
                                                },
                                                "broadcast": null
                                            }
                                        }
                                    },
                                    "Broadcast notification": {
                                        "summary": "Broadcast Notification Response",
                                        "value": {
                                            "message": "Broadcast notification sent successfully. It will be delivered to all elligible recipients.",
                                            "data": {
                                                "notification": null,
                                                "broadcast": {
                                                    "id": 42069,
                                                    "target": {
                                                        "channel": "announcements",
                                                        "topic": "product",
                                                        "event": "new_feature"
                                                    },
                                                    "payload": {
                                                        "title": "Big News!",
                                                        "message": "We just launched a new feature."
                                                    },
                                                    "created_at": "2025-08-11T20:54:25.039856+05:30",
                                                    "updated_at": "2025-08-11T20:54:25.039856+05:30"
                                                }
                                            }
                                        }
                                    },
                                    "Direct notification rejected": {
                                        "summary": "Rejected Notification Based on Preferences",
                                        "value": {
                                            "message": "No notification was sent. Recipient's preferences do not allow delivery.",
                                            "data": {
                                                "notification": null,
                                                "broadcast": null
                                            }
                                        }
                                    },
                                    "Direct notification with email": {
                                        "summary": "Direct Notification Fanned Out to Inbox + Email",
                                        "value": {
                                            "message": "Direct notification sent successfully to recipient recipient_123.",
                                            "data": {
                                                "notification": {
                                                    "id": 42070,
                                                    "recipient_id": "recipient_123",
                                                    "payload": {
                                                        "title": "Your daily digest is ready."
                                                    },
                                                    "broadcast_id": null,
                                                    "target": {
                                                        "channel": "digest",
                                                        "topic": "none",
                                                        "event": "sent"
                                                    },
                                                    "read": false,
                                                    "opened": false,
                                                    "created_at": "2025-11-07T05:31:56Z",
                                                    "updated_at": "2025-11-07T05:31:56Z"
                                                },
                                                "broadcast": null,
                                                "deliveries": [
                                                    {
                                                        "medium": "email",
                                                        "status": "pending",
                                                        "address": "alice@example.com",
                                                        "created_at": "2025-11-07T05:31:56Z",
                                                        "updated_at": "2025-11-07T05:31:56Z"
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid send. **Only projects with strict targets enabled** return the uncataloged-target error — with the setting off (the default) an uncataloged target is accepted. The check runs per medium, against the mediums the send actually asks for, and applies to direct sends and broadcasts alike. It tests for the **existence** of a catalog entry, not whether it is enabled: an entry with `default_enabled: false` is a real state (defined, switched off) and a send to it is accepted.",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "TargetNotCataloged": {
                                        "summary": "Target not in the project's catalog (strict targets on)",
                                        "value": {
                                            "message": "Target billing/none/invoice_paid is not in this project's catalog for the in_app medium. Create a project preference for it before sending, or turn off strict targets in your project settings."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuthWithAPIKeyWithFullAcessScope": []
                    }
                ]
            }
        },
...
        "/recipients": {
            "post": {
                "summary": "Create a recipient",
                "operationId": "createRecipient",
                "tags": [
                    "Recipients"
                ],
                "description": "Create a new recipient.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecipientPayload"
                            }
                        }
                    }
                },
                "x-codeSamples": [
                    {
                        "lang": "curl",
                        "label": "cURL",
                        "source": "curl -X POST https://api.bodhveda.com/recipients \\\n  -H \"Authorization: Bearer bv_xxxxxxxxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"id\": \"recipient_123\",\n    \"name\": \"Alice Johnson\"\n  }' "
                    },
                    {
                        "lang": "javascript",
                        "label": "JavaScript",
                        "source": "import { Bodhveda } from \"@bodhveda/js\";\n\nconst bodhveda = new Bodhveda(\"bv_xxxxxxxxx\");\n\nconst recipient = await bodhveda.recipients.create({\n    id: \"recipient_123\",\n    name: \"Alice Johnson\",\n});"
                    },
                    {
                        "lang": "go",
                        "label": "Go",
                        "source": "import (\n    \"context\"\n\n    bodhveda \"github.com/MudgalLabs/bodhveda/sdk/go\"\n)\n\nctx := context.Background()\nclient := bodhveda.NewClient(\"bv_xxxxxxxxx\", nil)\n\nname := \"Alice Johnson\"\nrecipient, _ := client.Recipients.Create(ctx, &bodhveda.CreateRecipientRequest{\n    ID:   \"recipient_123\",\n    Name: &name,\n})"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Recipient created successfully",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Recipient created": {
                                        "summary": "Recipient Created Response",
                                        "value": {
                                            "message": "Recipient created",
                                            "data": {
                                                "id": "recipient_123",
                                                "name": "John Doe",
                                                "created_at": "2025-11-07T05:31:56Z",
                                                "updated_at": "2025-11-07T05:31:56Z"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Recipient already exists",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Conflict": {
                                        "summary": "Recipient Already Exists",
                                        "value": {
                                            "message": "Recipient already exists"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuthWithAPIKeyWithFullAcessScope": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuthWithAPIKeyWithFullAcessScope": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "APIKey",
                "description": "Bearer authentication header of the form `Bearer bv_xxxxxxxxx`, where `bv_xxxxxxxxx` is your **API Key** with `Full access` scope."
            },
            "BearerAuthWithAPIKeyWithEitherScope": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "APIKey",
                "description": "Bearer authentication header of the form `Bearer bv_xxxxxxxxx`, where `bv_xxxxxxxxx` is your **API Key** with `Full access` or `Recipient acess` scope."
            }
        },
        "schemas": {
            "SendNotificationPayload": {
                "type": "object",
                "properties": {
                    "recipient_id": {
                        "type": "string",
                        "nullable": true,
                        "description": "The `id` of a recipient.\n- If null → this is treated as a **broadcast** notification.\n- If set → this is a **direct** notification to that recipient. \n\n 💡 If no recipient exists with the `recipient_id`, Bodhveda will create a new recipient with that `recipient_id`."
                    },
                    "target": {
                        "$ref": "#/components/schemas/Target"
                    },
                    "payload": {
                        "type": "object",
                        "description": "Arbitrary JSON payload for the notification. This is the **in-app** content block — its presence is what makes the notification eligible for in-app delivery.\n\n**Optional on a direct send.** Omit it to deliver on email *without* creating an in-app notification (an *email-only* send) — the `email` block then stands alone. An explicit `null` counts as omitted.\n\n**Required on a broadcast**, which is in-app only.\n\n⚠️ A send must carry at least one content block. Omitting both `payload` and `email` is a `400` — such a send names no medium.",
                        "nullable": true
                    },
                    "email": {
                        "$ref": "#/components/schemas/EmailContent"
                    }
                }
            },
...
            "CreateRecipientPayload": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique identifier from your system. \n\n This cannot be changed after creation. So use a stable identifier that will not change over time."
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the recipient."
                    }
                },
                "required": [
                    "id"
                ]
            },
...
        }
    }
}